Create script.py
So far, this can only get and print the directory address. For renaming, the script would need to parse through the last couple parts of the file path name and append it to a string. When this is done, the name of of the file will be changed to the string.